+++ apps/constellation/constellation.cpp
@@ -17,7 +17,7 @@
 //------------------------------------------------------------------------------
 
 #include "constellation.hpp"
-#include "core/bloom_filter.hpp"
+#include "core/bloom_filter_interface.hpp"
 #include "dkg/dkg_service.hpp"
 #include "health_check_http_module.hpp"
 #include "http/middleware/allow_origin.hpp"
