+++ libs/ledger/include/ledger/chain/main_chain.hpp
@@ -17,7 +17,7 @@
 //
 //------------------------------------------------------------------------------
 
-#include "core/bloom_filter.hpp"
+#include "core/bloom_filter_interface.hpp"
 #include "core/byte_array/byte_array.hpp"
 #include "core/byte_array/decoders.hpp"
 #include "core/mutex.hpp"
